Write spinnaker endpoints to shared profile
This is also a prereq to setting up settings.js in non-default deployments
