+++ halyard-config/src/main/java/com/netflix/spinnaker/halyard/config/model/v1/node/DeploymentEnvironment.java
@@ -66,6 +66,6 @@ public class DeploymentEnvironment extends Node {
     }
   }
 
-  DeploymentType type;
-  String accountName;
+  private DeploymentType type = DeploymentType.LocalhostDebian;
+  private String accountName;
 }
