+++ halyard-deploy/src/main/java/com/netflix/spinnaker/halyard/deploy/services/v1/GenerateService.java
@@ -1,0 +1,167 @@
+/*
+ * Copyright 2017 Google, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License")
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.netflix.spinnaker.halyard.deploy.services.v1;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.netflix.spinnaker.halyard.config.config.v1.AtomicFileWriter;
+import com.netflix.spinnaker.halyard.config.errors.v1.HalconfigException;
+import com.netflix.spinnaker.halyard.config.model.v1.node.DeploymentConfiguration;
+import com.netflix.spinnaker.halyard.config.model.v1.node.NodeFilter;
+import com.netflix.spinnaker.halyard.config.model.v1.problem.ProblemBuilder;
+import com.netflix.spinnaker.halyard.config.services.v1.DeploymentService;
+import com.netflix.spinnaker.halyard.config.spinnaker.v1.SpinnakerEndpoints;
+import com.netflix.spinnaker.halyard.config.spinnaker.v1.component.SpinnakerComponent;
+import com.netflix.spinnaker.halyard.deploy.deployment.v1.Deployment;
+import com.netflix.spinnaker.halyard.deploy.deployment.v1.DeploymentFactory;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.tomcat.util.http.fileupload.FileUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import org.yaml.snakeyaml.Yaml;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+
+import static com.netflix.spinnaker.halyard.config.model.v1.problem.Problem.Severity;
+
+@Component
+@Slf4j
+public class GenerateService {
+  @Autowired
+  String spinnakerOutputPath;
+
+  @Autowired
+  String halconfigDirectory;
+
+  @Autowired
+  DeploymentService deploymentService;
+
+  @Autowired
+  DeploymentFactory deploymentFactory;
+
+  @Autowired
+  String halconfigPath;
+
+  @Autowired
+  Yaml yamlParser;
+
+  @Autowired
+  ObjectMapper objectMapper;
+
+  @Autowired(required = false)
+  List<SpinnakerComponent> spinnakerComponents = new ArrayList<>();
+
+  void atomicWrite(Path path, String contents) {
+    AtomicFileWriter writer = null;
+    try {
+      writer = new AtomicFileWriter(path);
+      writer.write(contents);
+      writer.commit();
+    } catch (IOException ioe) {
+      ioe.printStackTrace();
+      throw new HalconfigException(
+          new ProblemBuilder(Severity.FATAL,
+              "Failed to write config for profile " + path.toFile().getName() + ": " + ioe
+                  .getMessage()).build()
+      );
+    } finally {
+      if (writer != null) {
+        writer.close();
+      }
+    }
+  }
+
+  public String yamlToString(Object yaml) {
+    return yamlParser.dump(objectMapper.convertValue(yaml, Map.class));
+  }
+
+  /**
+   * Generate config for a given deployment.
+   *
+   * This involves a few steps:
+   *
+   *   1. Clear out old config generated in a prior run.
+   *   2. Determine what the deployment footprint looks like to provide endpoint information for each service.
+   *   3. Generate configuration using the halconfig as the source of truth.
+   *   4. Copy custom profiles from the specified deployment over to the new deployment.
+   *
+   * @param nodeFilter A filter that specifies the deployment to use.
+   */
+  public void generateConfig(NodeFilter nodeFilter) {
+    DeploymentConfiguration deploymentConfiguration = deploymentService.getDeploymentConfiguration(nodeFilter);
+    String deploymentName = deploymentConfiguration.getName();
+
+    log.info("Generating config from \"" + halconfigPath + "\" with deploymentName \"" + deploymentName + "\"");
+
+    File spinnakerOutput = new File(spinnakerOutputPath);
+
+    // Step 1.
+    try {
+      FileUtils.deleteDirectory(spinnakerOutput);
+    } catch (IOException e) {
+      throw new HalconfigException(
+          new ProblemBuilder(Severity.FATAL, "Unable to clear old spinnaker config: " + e.getMessage() + ".").build());
+    }
+
+    if (!spinnakerOutput.mkdirs()) {
+      throw new HalconfigException(
+          new ProblemBuilder(Severity.FATAL, "Unable to create new spinnaker config directory \"" + spinnakerOutputPath + "\".").build());
+    }
+
+    // Step 2.
+    AtomicFileWriter writer = null;
+    Deployment deployment = deploymentFactory.create(deploymentConfiguration);
+    FileSystem defaultFileSystem = FileSystems.getDefault();
+    Path path = defaultFileSystem.getPath(spinnakerOutputPath, "spinnaker.yml");
+    log.info("Writing spinnaker endpoints");
+    SpinnakerEndpoints endpoints = deployment.getEndpoints();
+    atomicWrite(path, yamlToString(deployment.getEndpoints()));
+
+    // Step 3.
+    for (SpinnakerComponent component : spinnakerComponents) {
+      path = defaultFileSystem.getPath(spinnakerOutputPath, component.getConfigFileName());
+
+      log.info("Writing " + component.getComponentName() + " profile");
+      atomicWrite(path, component.getFullConfig(nodeFilter, endpoints));
+    }
+
+    // Step 4.
+    Path userProfilePath = Paths.get(halconfigDirectory, deploymentName);
+
+    if (Files.isDirectory(userProfilePath)) {
+      File[] files = userProfilePath.toFile().listFiles();
+      if (files == null) {
+        files = new File[0];
+      }
+
+      Arrays.stream(files).forEach(f -> {
+        try {
+          Files.copy(f.toPath(), Paths.get(spinnakerOutput.toString(), f.getName()), StandardCopyOption.REPLACE_EXISTING);
+        } catch (IOException e) {
+          throw new HalconfigException(
+              new ProblemBuilder(Severity.FATAL, "Unable to copy profile \"" + f.getName() + "\": " + e.getMessage() + ".").build()
+          );
+        }
+      });
+    }
+  }
+}
